Add support for StringData in secrets
This is identical to the PR made yesterday, except for the reverse situation in secrets. Secrets automatically assume all fields in data are base64-encoded, and now contain a new field, StringData, which allows a secret to contain non-base64-encoded data. This PR adds support for replicating StringData fields in secrets.